client can set g and q and enter their private key. they key can be sent to the server after moding it.
the server receives g and q.
it also receive the private key moded.
server ask user for their key and mod it.

But needs a case to handle g and q by comparing whether the first character is -1 then set g and q with a confirmation message back to the client server